requirements.txt
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==2.2.5
django-modeltranslation==0.17.1
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2-binary==2.8.6
requests==2.25.1
six==1.16.0
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
